Rename firebase hash to match the other files
