+++ lib/firebase-hash.js
